// generated/src/app/api-models/nav-win.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class NavWin {
  winId: string;
  pbname: string;
  pwinId: string;
  cwinId: string;
  sheet: string;
  menu: string;
  dbmaint: string;
  maxOpen: number;
  bitmap: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  dispOrder: number;
  notesCname: string;
  notesMembContext: string;
  notesGroupContext: string;
  notesProvContext: string;
  languageId: number;
}